third_party/blink/web_tests/fast/events/drag-and-drop-set-drag-data-arguments-expected.txt

Tests correct behavior of event.dataTransfer.setData with wrong number of arguments

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS event.dataTransfer.setData() threw exception TypeError: Failed to execute 'setData' on 'DataTransfer': 2 arguments required, but only 0 present..
PASS event.dataTransfer.setData("Text") threw exception TypeError: Failed to execute 'setData' on 'DataTransfer': 2 arguments required, but only 1 present..
PASS successfullyParsed is true

TEST COMPLETE

Drag me
